requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==2.2.9
django-allauth==0.41.0
django-contrib-comments==1.9.2
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-extensions==2.2.6
django-lifecycle==0.7.1
django-markdownx==3.0.1
idna==2.8
Markdown==3.2
martor==1.4.6
oauthlib==3.1.0
Pillow==7.0.0
pip-tools==4.4.0
psycopg2-binary==2.8.4
Pygments==2.5.2
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
urllib3==1.25.8
